feat: CNS wip - calculate in 0.5 - 1.6 range without half-time

+++ src/common/global_types.rs
@@ -6,4 +6,4 @@ pub type GradientFactor = u8;
 pub type GradientFactors = (u8, u8);
 pub type MbarPressure = u16;
 pub type AscentRatePerMinute = f64;
-pub type CNSPercent = u8;
+pub type CNSPercent = f64;
